requirements.txt
APScheduler==3.6.3
attrs==20.3.0
bravado==10.4.1
bravado-core==5.17.0
Bybit==0.2.11
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpointer==2.1
jsonref==0.2
jsonschema==3.2.0
MarkupSafe==1.1.1
monotonic==1.5
msgpack==1.0.2
msgpack-python==0.5.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-telegram-bot==13.4.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rfc3987==1.3.8
simplejson==3.17.2
six==1.15.0
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.24.2
webcolors==1.11.1
Werkzeug==1.0.1
